requirements.txt
Django>=1.11,<2.0
django-allauth==0.34.0
django-apptemplates==1.4

PyYAML
django-pipeline>=1.6.13,<1.7
# Fork to cope with newer pipeline
mysociety-django-pipeline-compass-rubygem
python-memcached
django-tastypie>=0.14.0

# Scrapers
beautifulsoup4
requests_cache

-e git+https://github.com/mysociety/sayit#egg=django-sayit
